Page store: long keys

+++ h2/src/main/org/h2/index/RowIndex.java
@@ -20,9 +20,9 @@ public interface RowIndex extends Index {
      * Get the row with the given key.
      *
      * @param session the session
-     * @param key the position
+     * @param key the unique key
      * @return the row
      */
-    Row getRow(Session session, int key) throws SQLException;
+    Row getRow(Session session, long key) throws SQLException;
 
 }
